update DefaultActiveAccountMonitor after move
